fix: 일기 썸네일 이미지 조건 검사에 "" 도 거르도록 조건 추가

+++ src/main/java/org/lxdproject/lxd/diary/service/DiaryService.java
@@ -173,7 +173,7 @@ public class DiaryService {
             throw new DiaryHandler(ErrorStatus.FORBIDDEN_DIARY_UPDATE);
         }
 
-        if (diary.getThumbImg() != null && !diary.getThumbImg().equals(request.getThumbImg())) {
+        if (diary.getThumbImg() != null && !diary.getThumbImg().isBlank() && !diary.getThumbImg().equals(request.getThumbImg())) {
             s3FileService.deleteImage(diary.getThumbImg());
         }
